% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyDiannFiltering}
\alias{ApplyDiannFiltering}
\title{Apply DIA-NN-specific filtering}
\usage{
ApplyDiannFiltering(
  data,
  qvalueFilter = TRUE,
  pepFilter = FALSE,
  pepThreshold = 0.01,
  minPeptides = 0
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{qvalueFilter}{Filter proteins by Q-value < 0.01}

\item{pepFilter}{Filter proteins by PEP threshold}

\item{pepThreshold}{PEP threshold}

\item{minPeptides}{Minimum peptide/precursor evidence count}
}
\value{
Filtered data matrix
}
\description{
Apply DIA-NN-specific filtering
}
