% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyDiannPhosphoFiltering}
\alias{ApplyDiannPhosphoFiltering}
\title{Apply DIA-NN phosphoproteomics filtering}
\usage{
ApplyDiannPhosphoFiltering(data, qvalueFilter = TRUE, minPeptides = 0)
}
\arguments{
\item{data}{Data matrix to filter}

\item{qvalueFilter}{Filter by Q-value < 0.01}

\item{minPeptides}{Minimum peptide/precursor evidence count}
}
\value{
List with filtered data and statistics
}
\description{
Apply DIA-NN phosphoproteomics filtering
}
