% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyFormatSpecificFiltering}
\alias{ApplyFormatSpecificFiltering}
\title{Apply format-specific filtering based on data format
Called before general filtering to apply MaxQuant/DIA-NN/FragPipe/Spectronaut specific filters}
\usage{
ApplyFormatSpecificFiltering(dataName, dataFormat, formatOpts = list())
}
\arguments{
\item{dataName}{Dataset name}

\item{dataFormat}{Format type (maxquant, diann, fragpipe, spectronaut)}

\item{formatOpts}{List of format-specific options}
}
\description{
Apply format-specific filtering based on data format
Called before general filtering to apply MaxQuant/DIA-NN/FragPipe/Spectronaut specific filters
}
