% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyFragpipeFiltering}
\alias{ApplyFragpipeFiltering}
\title{Apply FragPipe-specific filtering}
\usage{
ApplyFragpipeFiltering(
  data,
  removeContaminants = TRUE,
  minProb = 0.99,
  minPeptides = 2
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{removeContaminants}{Remove contaminant proteins}

\item{minProb}{Minimum protein probability threshold}

\item{minPeptides}{Minimum combined total peptides}
}
\value{
Filtered data matrix
}
\description{
Apply FragPipe-specific filtering
}
