% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyFragpipePhosphoFiltering}
\alias{ApplyFragpipePhosphoFiltering}
\title{Apply FragPipe phosphoproteomics filtering}
\usage{
ApplyFragpipePhosphoFiltering(
  data,
  removeContaminants = TRUE,
  minProb = 0,
  minPeptides = 0
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{removeContaminants}{Remove contaminant proteins}

\item{minProb}{Minimum localization probability (0-1)}

\item{minPeptides}{Minimum number of peptides}
}
\value{
List with filtered data and statistics
}
\description{
Apply FragPipe phosphoproteomics filtering
}
