% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyMaxQuantFiltering}
\alias{ApplyMaxQuantFiltering}
\title{Apply MaxQuant-specific filtering}
\usage{
ApplyMaxQuantFiltering(
  data,
  removeContaminants = TRUE,
  removeDecoys = TRUE,
  removeOnlyBySite = TRUE,
  minPeptides = 2
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{removeContaminants}{Remove contaminant proteins marked with "+"}

\item{removeDecoys}{Remove decoy/reverse proteins marked with "+"}

\item{minPeptides}{Minimum number of peptides required per protein}
}
\value{
Filtered data matrix
}
\description{
Apply MaxQuant-specific filtering
}
