% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplyMaxQuantPhosphoFiltering}
\alias{ApplyMaxQuantPhosphoFiltering}
\title{Apply MaxQuant phosphoproteomics filtering}
\usage{
ApplyMaxQuantPhosphoFiltering(
  data,
  removeContaminants = TRUE,
  removeDecoys = TRUE,
  minPeptides = 0
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{removeContaminants}{Remove contaminant proteins}

\item{removeDecoys}{Remove decoy/reverse proteins}

\item{minPeptides}{Minimum number of peptides (not typically used for phospho)}
}
\value{
List with filtered data and statistics
}
\description{
Apply MaxQuant phosphoproteomics filtering
}
