% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplySampleNormalization}
\alias{ApplySampleNormalization}
\title{Apply sample normalization (row-wise normalization)}
\usage{
ApplySampleNormalization(data, method = "none", sampleNormParam = "__manual__")
}
\arguments{
\item{data}{Matrix with samples as columns and features as rows}

\item{method}{Sample normalization method: "none", "SumNorm", "MedianNorm", "ProbNorm", "SpecNorm"}
}
\value{
Normalized data matrix
}
\description{
Apply sample normalization (row-wise normalization)
}
