% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{ApplySpectronautFiltering}
\alias{ApplySpectronautFiltering}
\title{Apply Spectronaut-specific filtering}
\usage{
ApplySpectronautFiltering(
  data,
  removeContaminants = TRUE,
  qvalueFilter = TRUE,
  qvalueThreshold = 0.01,
  minPeptides = 0
)
}
\arguments{
\item{data}{Data matrix to filter}

\item{removeContaminants}{Remove contaminant proteins}
}
\value{
Filtered data matrix
}
\description{
Apply Spectronaut-specific filtering
}
