% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{CalculateFeatureRanking}
\alias{CalculateFeatureRanking}
\title{Calculate feature ranking for all features}
\usage{
CalculateFeatureRanking(dataName = "", clust.num = 5)
}
\arguments{
\item{dataName}{Dataset name}

\item{clust.num}{Number of clusters for k-means}
}
\description{
Calculate feature ranking for all features
}
