% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{CheckMetaIntegrity}
\alias{CheckMetaIntegrity}
\title{Check the integrity and consistency of metadata for meta-analysis}
\usage{
CheckMetaIntegrity()
}
\value{
Returns 1 to indicate successful completion of checks and updates.
}
\description{
This function checks the integrity and consistency of metadata across
different datasets intended for meta-analysis. It ensures that metadata
is properly aligned, consistent, and categorized as discrete or continuous.
}
\examples{
CheckMetaIntegrity()

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
