% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{CleanNumber}
\alias{CleanNumber}
\title{Replace infinite numbers}
\usage{
CleanNumber(bdata)
}
\arguments{
\item{bdata}{Input matrix to clean numbers}
}
\description{
Replace -Inf, Inf to 99999 and -99999
}
\author{
Jeff Xia\email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
