% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{ClearSetCache}
\alias{ClearSetCache}
\title{Clear the in-memory cache for readSet/saveSet}
\usage{
ClearSetCache(verbose = TRUE)
}
\arguments{
\item{verbose}{Logical. If TRUE, print cache size before clearing.}
}
\description{
Removes all cached objects from memory. Use this if you need to force
re-reading from disk files.
}
