% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ComputeAverageCurve}
\alias{ComputeAverageCurve}
\title{Compute average ROC curve}
\usage{
ComputeAverageCurve(perf, avg.method)
}
\arguments{
\item{perf}{Input performance object}

\item{avg.method}{Method name}
}
\description{
Compute the average ROC curve
}
