% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ComputeHighLow}
\alias{ComputeHighLow}
\title{Compute the 95 percent interval for threshold ROC}
\usage{
ComputeHighLow(perf)
}
\arguments{
\item{perf}{Input performance object}
}
\description{
Computes the 95 percent interval only for the y-axis
}
