% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_network_utils.R
\name{ComputeSubnetStats}
\alias{ComputeSubnetStats}
\title{ComputeSubnetStats}
\usage{
ComputeSubnetStats(comps)
}
\arguments{
\item{comps}{List of igraph components}
}
\value{
Data frame with Node, Edge, Query counts for each component
}
\description{
Helper function to compute basic statistics for network components
This is needed by DecomposeGraph which is called during PPI network building
}
