% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{ContainNewSamples}
\alias{ContainNewSamples}
\title{Check if contains new samples}
\usage{
ContainNewSamples(dataName = "")
}
\arguments{
\item{dataName}{Input dataset name}
}
\description{
Check if contains new samples
}
