% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R, R/covariate_utils.R
\name{CovariateScatter.Anal}
\alias{CovariateScatter.Anal}
\title{Perform Covariate Scatter Analysis}
\usage{
CovariateScatter.Anal(
  dataName,
  imgName = "NA",
  imgFormat = "png",
  analysis.var,
  ref = NULL,
  block = "NA",
  thresh = 0.05,
  pval.selection = "fdr",
  contrast.cls = "anova",
  useBatchCorrected = TRUE
)

CovariateScatter.Anal(
  dataName,
  imgName = "NA",
  imgFormat = "png",
  analysis.var,
  ref = NULL,
  block = "NA",
  thresh = 0.05,
  pval.selection = "fdr",
  contrast.cls = "anova",
  useBatchCorrected = TRUE
)
}
\arguments{
\item{dataName}{A character string specifying the name of the dataset.}

\item{imgName}{A character string specifying the name of the image to save.}

\item{imgFormat}{The format of the image to create (e.g., "png", "jpeg").}

\item{analysis.var}{The analysis variable.}

\item{ref}{The reference variable.}

\item{block}{The blocking variable.}

\item{thresh}{The significance threshold.}

\item{pval.selection}{The method for p-value selection ("fdr" or "raw").}

\item{contrast.cls}{The contrast class ("anova" or "t").}
}
\value{
A vector with c(sig.num, nonSig) - number of significant and non-significant features.

Result is saved in analSet object.
}
\description{
This function performs a covariate scatter analysis to visualize the effect of 
covariate adjustment on p-values. It generates a JSON file for interactive visualization.

This function performs a covariate scatter analysis based on the specified parameters.
}
\examples{
\dontrun{
CovariateScatter.Anal(dataName = "estrogen.txt", analysis.var = "variableA",
                      imgName = "scatter_plot", imgFormat = "png",
                      thresh = 0.05, pval.selection = "fdr",
                      contrast.cls = "anova")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
