% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{createCVset}
\alias{createCVset}
\title{Separate data set using k-fold cross validation (CV)}
\usage{
createCVset(groupN, kfold, rseed)
}
\arguments{
\item{groupN}{Input group size}

\item{kfold}{Input the number of folds}

\item{rseed}{Input the random seed}
}
\description{
Separate data set with k-fold CV
}
