% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_utils.R
\name{doEntrez2SymbolMappingPhospho}
\alias{doEntrez2SymbolMappingPhospho}
\title{Wrapper for doEntrez2SymbolMapping that handles phosphosites}
\usage{
doEntrez2SymbolMappingPhospho(id.vec, data.org = "NA", data.idType = "NA")
}
\arguments{
\item{id.vec}{Character vector - phosphosite IDs if phospho data, else entrez IDs}

\item{data.org}{Organism}

\item{data.idType}{ID type}
}
\value{
Character vector of gene symbols
}
\description{
Wrapper for doEntrez2SymbolMapping that handles phosphosites
}
