% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_utils.R
\name{doEntrezIDAnotPhospho}
\alias{doEntrezIDAnotPhospho}
\title{Wrapper for doEntrezIDAnot that handles phosphosites}
\usage{
doEntrezIDAnotPhospho(id.vec, data.org = "NA", data.idType = "NA")
}
\arguments{
\item{id.vec}{Character vector - phosphosite IDs if phospho data, else entrez IDs}

\item{data.org}{Organism}

\item{data.idType}{ID type}
}
\value{
Data frame with gene_id, symbol, name
}
\description{
Wrapper for doEntrezIDAnot that handles phosphosites
}
