% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{doLogisticRegMdl}
\alias{doLogisticRegMdl}
\title{Develop a Logistic Regression Model with k-fold CV}
\usage{
doLogisticRegMdl(dataName = "", x.train, y.train, x.test, y.test)
}
\arguments{
\item{dataName}{Input dataset name}

\item{x.train}{Training X}

\item{y.train}{Training Y}

\item{x.test}{Test X}

\item{y.test}{Test Y}
}
\description{
Develop LR Model with CV
}
