% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{.buildKinaseSubstrateNetwork}
\alias{.buildKinaseSubstrateNetwork}
\title{Build kinase-substrate network structure}
\usage{
.buildKinaseSubstrateNetwork(
  sig_kinases,
  ks_db,
  sig_sites,
  all_sites,
  dataSet,
  paramSet
)
}
\description{
Build kinase-substrate network structure
}
\keyword{internal}
