% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{.correctPhosphoByProteinAbundance}
\alias{.correctPhosphoByProteinAbundance}
\title{Correct phosphosite intensities by protein abundance}
\usage{
.correctPhosphoByProteinAbundance(phospho_data, protein_ref)
}
\arguments{
\item{phospho_data}{Matrix of phosphosite intensities (log2-transformed), sites × samples}

\item{protein_ref}{Matrix of protein abundances (log2-transformed), proteins × samples}
}
\value{
Corrected phosphosite matrix with protein abundance subtracted
}
\description{
Correct phosphosite intensities by protein abundance
}
