% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.countSiteOverlaps}
\alias{.countSiteOverlaps}
\title{Count overlaps between phosphosite IDs and kinase substrate list}
\usage{
.countSiteOverlaps(site_ids, substrates)
}
\arguments{
\item{site_ids}{Vector of phosphosite IDs (format: PROTEIN_RES_POS like P12345_S_123)}

\item{substrates}{Vector of substrates (format: "PROTEIN;RES_POS;" like "TP53;S15;")}
}
\value{
Number of matching sites
}
\description{
Count overlaps between phosphosite IDs and kinase substrate list
}
