% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{.do.CVTest.LRmodel}
\alias{.do.CVTest.LRmodel}
\title{Calculate ROC performance with CV for LR model}
\usage{
.do.CVTest.LRmodel(data.in, fmla.in, kfold = 10, run.stepwise = FALSE)
}
\arguments{
\item{data.in}{Input matrix of data}

\item{fmla.in}{Input for generalized linear model}

\item{kfold}{Numeric}

\item{run.stepwise}{Logistic Regression}
}
\description{
Internal function for LR cross-validation
}
