% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.enrichmentFromCuratedDB}
\alias{.enrichmentFromCuratedDB}
\title{Enrichment from curated kinase-substrate database}
\usage{
.enrichmentFromCuratedDB(sig_sites, all_sites, ks_db)
}
\description{
Enrichment from curated kinase-substrate database
}
