% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.enrichmentFromListDB}
\alias{.enrichmentFromListDB}
\title{Enrichment from list-based kinase-substrate database
Database format: list(kinase = c("PROTEIN;RESIDUE_POS;", ...))}
\usage{
.enrichmentFromListDB(sig_sites, all_sites, ks_db)
}
\description{
Enrichment from list-based kinase-substrate database
Database format: list(kinase = c("PROTEIN;RESIDUE_POS;", ...))
}
