% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.enrichmentFromMotifs}
\alias{.enrichmentFromMotifs}
\title{Enrichment from motif-based kinase mapping}
\usage{
.enrichmentFromMotifs(sig_sites, all_sites, motif_db, dataSet)
}
\description{
Enrichment from motif-based kinase mapping
}
