% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{.exportKinaseNetworkToJSON}
\alias{.exportKinaseNetworkToJSON}
\title{Export kinase-substrate network to JSON format (same as enrichment network)}
\usage{
.exportKinaseNetworkToJSON(network_data, netNm, paramSet)
}
\description{
Export kinase-substrate network to JSON format (same as enrichment network)
}
\keyword{internal}
