% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_network_utils.R
\name{.exportNetworkToJSON}
\alias{.exportNetworkToJSON}
\title{.exportNetworkToJSON}
\usage{
.exportNetworkToJSON(g, json.nm, net.nm)
}
\arguments{
\item{g}{igraph object}

\item{json.nm}{Output file name}

\item{net.nm}{Network name}
}
\description{
Internal function to convert igraph to JSON for cytoscape/sigma.js visualization
}
