% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{.generateKinaseNetworkTable}
\alias{.generateKinaseNetworkTable}
\title{Generate summary table for kinase-substrate network}
\usage{
.generateKinaseNetworkTable(network_data, netNm)
}
\description{
Generate summary table for kinase-substrate network
}
\keyword{internal}
