% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_utils.R
\name{.isPhosphoData}
\alias{.isPhosphoData}
\title{Check if data type is phosphoproteomics}
\usage{
.isPhosphoData()
}
\value{
Logical TRUE if phospho data
}
\description{
Check if data type is phosphoproteomics
}
