% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.loadKinaseSubstrateDB}
\alias{.loadKinaseSubstrateDB}
\title{Load kinase-substrate database}
\usage{
.loadKinaseSubstrateDB(database = "phosphositeplus", paramSet)
}
\arguments{
\item{database}{Database name}

\item{paramSet}{Parameter set with organism info}
}
\value{
Data frame with kinase-substrate relationships
}
\description{
Load kinase-substrate database
}
