% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{.matchSubstrateIDsToSites}
\alias{.matchSubstrateIDsToSites}
\title{Match kinase substrate IDs from database to user's phosphosite data
Substrate IDs from database are typically in format: "UNIPROT;SITE" (e.g., "Q80XQ2;S578")}
\usage{
.matchSubstrateIDsToSites(
  substrate_ids,
  site_info,
  all_site_info,
  sig_sites,
  all_sites
)
}
\description{
Match kinase substrate IDs from database to user's phosphosite data
Substrate IDs from database are typically in format: "UNIPROT;SITE" (e.g., "Q80XQ2;S578")
}
\keyword{internal}
