% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{.ov_signif_cols}
\alias{.ov_signif_cols}
\title{Execute heavy package function in isolated RSclient fork}
\usage{
.ov_signif_cols(dat, digits = getOption("ov.csv.signif", 6))
}
\arguments{
\item{func_body}{Function(input_data) to run in child}

\item{input_data}{List serialized via qs to child}

\item{packages}{Packages to load in child before execution}

\item{timeout}{Timeout in seconds}

\item{output_type}{"qs" for complex objects, "arrow" for data frames}
}
\value{
Result from child process
}
\description{
Execute heavy package function in isolated RSclient fork
}
