% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deanal_utils.R
\name{.paNormalizeProteinId}
\alias{.paNormalizeProteinId}
\title{Get protein-peptide mapping with DE results}
\usage{
.paNormalizeProteinId(x)
}
\arguments{
\item{dataName}{Dataset name}

\item{proteinID}{Protein ID to get peptides for}
}
\description{
Returns a data frame with protein and associated peptides with DE stats
}
