% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{.parsePhosphositeIDs}
\alias{.parsePhosphositeIDs}
\title{Parse phosphosite IDs into components}
\usage{
.parsePhosphositeIDs(site_ids)
}
\description{
Parse phosphosite IDs into components
}
\keyword{internal}
