% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_utils.R
\name{.performEnrichAnalysisPhospho}
\alias{.performEnrichAnalysisPhospho}
\title{Perform enrichment analysis for phosphosite data}
\usage{
.performEnrichAnalysisPhospho(
  dataSet,
  file.nm,
  fun.type,
  phosphosite.vec,
  vis.type
)
}
\arguments{
\item{dataSet}{Dataset object}

\item{file.nm}{Output file name}

\item{fun.type}{Function library type}

\item{phosphosite.vec}{Named vector of phosphosite IDs (names are symbols)}

\item{vis.type}{Visualization type}
}
\value{
Result code
}
\description{
Perform enrichment analysis for phosphosite data
}
