% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrichment_utils.R
\name{.performEnrichAnalysisPhosphoInternal}
\alias{.performEnrichAnalysisPhosphoInternal}
\title{Modified performEnrichAnalysis that returns phosphosite IDs in fun.anot
This is called internally by .performEnrichAnalysisPhospho}
\usage{
.performEnrichAnalysisPhosphoInternal(
  dataSet,
  file.nm,
  fun.type,
  ora.vec,
  vis.type
)
}
\description{
Modified performEnrichAnalysis that returns phosphosite IDs in fun.anot
This is called internally by .performEnrichAnalysisPhospho
}
