% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{.performKinaseEnrichmentTest}
\alias{.performKinaseEnrichmentTest}
\title{Perform kinase enrichment test using Fisher's exact test}
\usage{
.performKinaseEnrichmentTest(sig_sites, all_sites, ks_db, dataSet)
}
\arguments{
\item{sig_sites}{Significant phosphosites}

\item{all_sites}{All detected phosphosites}

\item{ks_db}{Kinase-substrate database (list format)}

\item{dataSet}{Dataset object with feature info}
}
\value{
Data frame with enrichment results
}
\description{
Perform kinase enrichment test using Fisher's exact test
}
