% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_arrow.R
\name{.prepare_df_with_rownames}
\alias{.prepare_df_with_rownames}
\title{Internal: Prepare data frame with row_names_id column}
\usage{
.prepare_df_with_rownames(obj)
}
\arguments{
\item{obj}{Matrix or data.frame}
}
\value{
Data frame with row_names_id as first column (if rownames exist)
}
\description{
Internal: Prepare data frame with row_names_id column
}
\keyword{internal}
