% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_data_utils.R
\name{.readDIANNProteome}
\alias{.readDIANNProteome}
\title{Read DIA-NN protein group matrix for phosphoproteomics normalization}
\usage{
.readDIANNProteome(proteinFile, phospho_samples)
}
\arguments{
\item{proteinFile}{Path to DIA-NN pg_matrix.tsv file}

\item{phospho_samples}{Character vector of sample names from phospho data}
}
\value{
Matrix of log2 protein abundances with proteins as rows, samples as columns
}
\description{
Read DIA-NN protein group matrix for phosphoproteomics normalization
}
