% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_data_utils.R
\name{.readMaxQuantProteome}
\alias{.readMaxQuantProteome}
\title{Read MaxQuant protein abundance data for phosphoproteomics normalization}
\usage{
.readMaxQuantProteome(proteinFile, phospho_samples)
}
\arguments{
\item{proteinFile}{Path to MaxQuant proteinGroups.txt file}

\item{phospho_samples}{Character vector of sample names from phospho data}
}
\value{
Matrix of protein abundances with proteins as rows, samples as columns
}
\description{
Read MaxQuant protein abundance data for phosphoproteomics normalization
}
