% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{EnableSetCache}
\alias{EnableSetCache}
\title{Enable in-memory caching for readSet/saveSet}
\usage{
EnableSetCache(enable = TRUE)
}
\arguments{
\item{enable}{Logical. TRUE to enable caching, FALSE to disable.}
}
\description{
When enabled, readSet will check memory cache before reading from disk,
and saveSet will update the cache. This can significantly speed up
repeated reads of the same objects (paramSet, analSet, etc.)
}
