% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{fast.write}
\alias{fast.write}
\title{Execute heavy package function in isolated RSclient fork}
\usage{
fast.write(dat, file, row.names = TRUE)
}
\arguments{
\item{func_body}{Function(input_data) to run in child}

\item{input_data}{List serialized via qs to child}

\item{packages}{Packages to load in child before execution}

\item{timeout}{Timeout in seconds}

\item{output_type}{"qs" for complex objects, "arrow" for data frames}
}
\value{
Result from child process
}
\description{
Execute heavy package function in isolated RSclient fork
}
