% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{FilteringData}
\alias{FilteringData}
\title{Filter Data by Count and Variance}
\usage{
FilteringData(nm, countOpt = "pct", count, var)
}
\arguments{
\item{nm}{The name of the dataset to be processed.}

\item{countOpt}{The option to use for count filtering. Can be "pct" (percentage) or "count" (absolute count).}

\item{count}{The count threshold for filtering.}

\item{var}{The variance threshold for filtering.}
}
\value{
The filtered dataset.
}
\description{
This function filters the data in a dataset by count and variance thresholds.
}
