% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{genLREquation}
\alias{genLREquation}
\title{Generate LR equation}
\usage{
genLREquation(coef.mdl)
}
\arguments{
\item{coef.mdl}{Model coefficients}
}
\description{
Generate equation string from LR coefficients
}
