% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.Accuracy}
\alias{Get.Accuracy}
\title{Calculate accuracy from confusion matrix}
\usage{
Get.Accuracy(cm)
}
\arguments{
\item{cm}{Confusion matrix}
}
\description{
Calculate accuracy from confusion matrix
}
