% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.pAUC}
\alias{Get.pAUC}
\title{Calculate partial area under ROC curve}
\usage{
Get.pAUC(x, y, focus, cutoff)
}
\arguments{
\item{x}{Input X}

\item{y}{Input Y}

\item{focus}{Method}

\item{cutoff}{Numeric}
}
\description{
Calculate partial AUC
}
