% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Get.VIP}
\alias{Get.VIP}
\title{Get VIP scores from PLS}
\usage{
Get.VIP(pls.obj, comp = 2)
}
\arguments{
\item{pls.obj}{PLS model object}

\item{comp}{Number of components}
}
\description{
Get VIP scores from PLS
}
