% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{GetAvailableKinaseDBs}
\alias{GetAvailableKinaseDBs}
\title{List available kinase databases for the current organism}
\usage{
GetAvailableKinaseDBs()
}
\value{
character vector of database names (includes "all" if >1)
}
\description{
List available kinase databases for the current organism
}
