% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetCIs}
\alias{GetCIs}
\title{Calculate confidence intervals}
\usage{
GetCIs(data, param = F)
}
\arguments{
\item{data}{Matrix of values (rows=runs, cols=models)}

\item{param}{Use parametric CI?}
}
\description{
Calculate confidence intervals
}
