% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetImpFeatureMat}
\alias{GetImpFeatureMat}
\title{Get important feature matrix}
\usage{
GetImpFeatureMat(dataName = "", feat.outp, bestFeatNum)
}
\arguments{
\item{dataName}{Input dataset name}

\item{feat.outp}{Feature output list}

\item{bestFeatNum}{Best feature number}
}
\description{
Get feature importance matrix from CV runs
}
