% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{GetKinaseEnrichmentResults}
\alias{GetKinaseEnrichmentResults}
\title{Get kinase enrichment results for display}
\usage{
GetKinaseEnrichmentResults()
}
\value{
Data frame with enrichment results
}
\description{
Get kinase enrichment results for display
}
