% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetLassoFreqs}
\alias{GetLassoFreqs}
\title{Compute lasso frequency}
\usage{
GetLassoFreqs(dataName = "")
}
\arguments{
\item{dataName}{Input dataset name}
}
\description{
Compute lasso frequency for feature selection
}
