% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetNewSampleNames}
\alias{GetNewSampleNames}
\title{Get new sample names}
\usage{
GetNewSampleNames(dataName = "")
}
\arguments{
\item{dataName}{Input dataset name}
}
\description{
Obtain sample names for new samples
}
