% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROC.coords}
\alias{GetROC.coords}
\title{Return ROC coordinates with confidence intervals}
\usage{
GetROC.coords(dataName = "", fld.nm, val, plot = TRUE, imgNm)
}
\arguments{
\item{dataName}{Input dataset name}

\item{fld.nm}{The kind of input coordinate}

\item{val}{The coordinates to look for}

\item{plot}{Logical, by default TRUE}

\item{imgNm}{Input the image name}
}
\description{
Return ROC coordinates with CI
}
