% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetROCTtestP}
\alias{GetROCTtestP}
\title{Calculate T-test p-values for ROC}
\usage{
GetROCTtestP(data, cls)
}
\arguments{
\item{data}{Feature matrix (samples × features)}

\item{cls}{Class labels}
}
\description{
Calculate T-test p-values for ROC
}
