% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_feattable_utils.R
\name{GetSigfeatures}
\alias{GetSigfeatures}
\title{Get Significant features from Analysis}
\usage{
GetSigfeatures(
  dataName = "",
  res.nm = "nm",
  p.lvl = 0.05,
  fc.lvl = 1,
  inx = 1,
  FDR = "true"
)
}
\arguments{
\item{dataName}{A character string specifying the name of the dataset.}

\item{res.nm}{A character string specifying the name of the output result table}

\item{p.lvl}{The significance threshold for p-values.}

\item{fc.lvl}{The fold change threshold.}

\item{inx}{The index for comparison (e.g., in case of multiple comparisons).}
}
\value{
A list of information including the filename, significant gene details, and counts.
}
\description{
This function retrieves significant features from the DE analysis based on the specified parameters.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
GetSigfeatures(dataName = "MyData", res.nm = "result_A", p.lvl = 0.05,
            fc.lvl = 1, inx = 1)
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
