% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{GetTrainTestSplitMat}
\alias{GetTrainTestSplitMat}
\title{Make random partitions}
\usage{
GetTrainTestSplitMat(y, propTraining = 2/3, nRuns = 30)
}
\arguments{
\item{y}{Input the data}

\item{propTraining}{By default set to 2/3}

\item{nRuns}{By default set to 30}
}
\description{
Make random partitions, returns matrices indicating
whether the observation is in train/test for each run
note: try to get a balanced sampling for each group (classification)
or each quantile (regression). This is very useful for unbalanced data
}
