% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_impute.R
\name{ImputeMissingVar}
\alias{ImputeMissingVar}
\title{Impute Missing Values in Variables}
\usage{
ImputeMissingVar(dataName = "", method = "min")
}
\arguments{
\item{dataName}{The name of the dataset to be processed.}

\item{method}{The imputation method to be used. Can be one of "exclude", "min", "colmin", "mean", "median", "knn_var", "knn_smp", "bpca", "ppca", "svdImpute".}
}
\description{
This function imputes missing values in variables (columns) of the dataset using various methods.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
ImputeMissingVar("data_filename", method = "min")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
