% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_general.R
\name{Init.Data}
\alias{Init.Data}
\title{Initialize resources for analysis}
\usage{
Init.Data(onWeb = T, dataPath = "data/", default.dpi = 72)
}
\arguments{
\item{onWeb}{whether the script is running in local or on web}
}
\description{
call this function before performing any analysis
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
