% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{logisticReg}
\alias{logisticReg}
\title{Logistic regression helper function}
\usage{
logisticReg(d.xy)
}
\arguments{
\item{d.xy}{Data frame with y and predictors}
}
\description{
Build logistic regression model
}
