% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{multi.stat}
\alias{multi.stat}
\title{Get multiple category statistics}
\usage{
multi.stat(pred, resp)
}
\arguments{
\item{pred}{Input predictions}

\item{resp}{Input responses}
}
\description{
Calculate sensitivity, specificity, PPV, NPV, etc.
}
