% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_gsea.R
\name{my.perform.gsea}
\alias{my.perform.gsea}
\title{Perform Gene Set Enrichment Analysis test on single or multiple gene expression matrices}
\usage{
my.perform.gsea(
  dataName,
  file.nm,
  fun.type,
  netNm,
  mType,
  selectedFactorInx = 1,
  mode = "multi",
  rankOpt = ""
)
}
\arguments{
\item{dataSetObj}{file name of the data, .txt format}
}
\description{
Perform Gene Set Enrichment Analysis test on single or multiple gene expression matrices
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
