% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Perform.Permut}
\alias{Perform.Permut}
\title{Perform permutation tests only for ROC Tester}
\usage{
Perform.Permut(dataName = "", perf.measure, perm.num, propTraining = 2/3)
}
\arguments{
\item{dataName}{Input the name of the dataset}

\item{perf.measure}{Performance measure ("auroc" or "accu")}

\item{perm.num}{Number of permutations}

\item{propTraining}{Fraction of samples for training}
}
\description{
Perform permutation tests for biomarker model
}
