% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Perform.UnivROC}
\alias{Perform.UnivROC}
\title{Perform Classical Univariate ROC}
\usage{
Perform.UnivROC(
  dataName = "",
  feat.nm,
  version,
  format = "png",
  dpi = 72,
  isAUC,
  isOpt,
  optMethod,
  isPartial,
  measure,
  cutoff,
  dataType = "default"
)
}
\arguments{
\item{dataName}{Input dataset name}

\item{feat.nm}{Input the feature name}

\item{version}{image version mark}

\item{format}{Select the image format}

\item{dpi}{Input the dpi}

\item{isAUC}{Logical, compute CI}

\item{isOpt}{Logical, show optimal cutoff}

\item{optMethod}{Optimal cutoff method}

\item{isPartial}{Logical, calculate partial ROC}

\item{measure}{Parameter to limit pROC}

\item{cutoff}{Threshold for pROC}
}
\description{
Perform Classical Univariate ROC analysis
}
