% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{PerformBatchCorrection}
\alias{PerformBatchCorrection}
\title{Perform Batch Correction}
\usage{
PerformBatchCorrection()
}
\value{
A list of datasets after batch correction.
}
\description{
This function performs batch correction using the ComBat method.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
PerformBatchCorrection()
}

}
\author{
Guangyan Zhou\email{guangyan.zhou@mail.mcgill.ca}
}
