% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PerformCV.test}
\alias{PerformCV.test}
\title{Perform MCCV for manually selected features}
\usage{
PerformCV.test(dataName = "", method, lvNum, propTraining = 2/3, nRuns = 100)
}
\arguments{
\item{dataName}{Input the name of the dataset}

\item{method}{Select the classification method}

\item{lvNum}{Input the number of latent variables}

\item{propTraining}{Proportion of samples for training}

\item{nRuns}{Number of MCCV runs}
}
\description{
MCCV for manually selected features (no additional feature selection)
}
