% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deanal_utils.R
\name{PerformDEAnal}
\alias{PerformDEAnal}
\title{Perform Differential Analysis}
\usage{
PerformDEAnal(
  dataName = "",
  anal.type = "default",
  par1 = NULL,
  par2 = NULL,
  nested.opt = "intonly",
  robustTrend = F
)
}
\arguments{
\item{dataName}{A character string specifying the name of the dataset.}

\item{anal.type}{The type of analysis to perform. Options: "default", "custom", "time", "reference", "nested".}

\item{par1}{Parameter 1, depending on the analysis type.}

\item{par2}{Parameter 2, depending on the analysis type.}

\item{nested.opt}{Option for nested analysis. Options: "intonly" (default), "all".}

\item{robustTrend}{Logical. If TRUE, use robust trend test.}
}
\value{
Results of the differential analysis.
}
\description{
This function performs differential analysis based on different types of comparisons.
}
\details{
default: all pair-wise comparison (A-B) + (B-C) + (A-C), custom: only compare two groups (A-C), time: compare consecutive groups (B-A) + (C-B), reference: all others against common reference (A-C) + (B-C), nested: (A-B)+(C-D)
}
\author{
Guangyan Zhou, \email{guangyan.zhou@mail.mcgill.ca}
}
