% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_network_utils.R
\name{PerformKinaseSubstrateNetworkView}
\alias{PerformKinaseSubstrateNetworkView}
\title{Generate kinase-substrate network visualization in enrichment network JSON format}
\usage{
PerformKinaseSubstrateNetworkView(
  dataName = "",
  netNm = "kinase_substrate_net",
  fdr_cutoff = 0.05
)
}
\arguments{
\item{dataName}{Dataset name}

\item{netNm}{Network name for output files}

\item{fdr_cutoff}{FDR cutoff for kinase filtering}
}
\value{
1 on success, 0 on failure
}
\description{
Generate kinase-substrate network visualization in enrichment network JSON format
}
