% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deanal_utils.R
\name{PerformLimmaDE}
\alias{PerformLimmaDE}
\title{Perform differential analysis using Limma method (meta-analysis)}
\usage{
PerformLimmaDE(dataName = "", grps, p.lvl, fc.lvl = NULL)
}
\arguments{
\item{dataName}{File name of data set.}

\item{grps}{Selected two groups for comparison}

\item{p.lvl}{P-value threshold}

\item{fc.lvl}{Fold-change threshold}
}
\description{
Detect number of DE features in individual data set for quality checking purposes
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
