% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deanal_utils.R
\name{PerformPeptideLevelDEAnal}
\alias{PerformPeptideLevelDEAnal}
\title{Perform peptide-level differential analysis}
\usage{
PerformPeptideLevelDEAnal(dataName = "")
}
\arguments{
\item{dataName}{Dataset name}
}
\description{
Performs DE analysis on peptide-level data after protein-level analysis is complete
}
