% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_data_utils.R
\name{phosphoCollapseLocal}
\alias{phosphoCollapseLocal}
\title{Collapse phosphosite-level data to protein-level}
\usage{
phosphoCollapseLocal(mat, id, stat, by = "max")
}
\description{
This function collapses multiple phosphosites to one value per protein/gene,
using a local collapse function. The site with the maximum absolute
statistic is chosen to represent each protein.
}
