% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{PlotDataProfile}
\alias{PlotDataProfile}
\title{Plot Data Profile}
\usage{
PlotDataProfile(dataName, type, boxplotName, pcaName)
}
\arguments{
\item{dataName}{The name of the dataset to be used for plotting.}

\item{type}{The type of data profile to plot (e.g., "boxplot" or "pca").}

\item{boxplotName}{The name of the file to save the box plot image.}

\item{pcaName}{The name of the file to save the PCA plot image.}
}
\description{
This function generates data profile plots for quality control, including box plots and PCA plots.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
PlotDataProfile("data_filename", "boxplot", "boxplot_image.png", "pca_image.png")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
