% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotDetailROC}
\alias{PlotDetailROC}
\title{Plot detailed ROC}
\usage{
PlotDetailROC(dataName = "", imgName, thresh, sp, se, dpi = 72, format = "png")
}
\arguments{
\item{dataName}{Input dataset name}

\item{imgName}{Input a name for the plot}

\item{thresh}{Input the threshold}

\item{sp}{Specificity}

\item{se}{Sensitivity}

\item{dpi}{Input the dpi}

\item{format}{Select the image format}
}
\description{
Plot detailed ROC with threshold
}
