% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{PlotKinaseEnrichment}
\alias{PlotKinaseEnrichment}
\title{Plot kinase enrichment bubble chart}
\usage{
PlotKinaseEnrichment(imgName, dpi = 96, format = "png", top_n = 20)
}
\arguments{
\item{imgName}{Image file name}

\item{dpi}{DPI for image}

\item{format}{Image format (png, pdf)}

\item{top_n}{Number of top kinases to show}
}
\value{
Image file name
}
\description{
Plot kinase enrichment bubble chart
}
