% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_qc_utils.R
\name{PlotLocalizationProbDist}
\alias{PlotLocalizationProbDist}
\title{Plot distribution of localization probabilities
Shows histogram with Class I (>0.75), II, and III sites}
\usage{
PlotLocalizationProbDist(fileName, imgNm, dpi = 72, format = "png")
}
\description{
Plot distribution of localization probabilities
Shows histogram with Class I (>0.75), II, and III sites
}
