% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariate_utils.R
\name{PlotMultiFacCmpdSummary}
\alias{PlotMultiFacCmpdSummary}
\title{Plot compound summary for multi-linear regression tool}
\usage{
PlotMultiFacCmpdSummary(
  dataName,
  imgName,
  name,
  id,
  meta,
  version,
  format = "png",
  dpi = 96,
  width = NA
)
}
\arguments{
\item{meta}{Input the metadata to visualize}

\item{version}{version}

\item{format}{Input the format of the image to create}

\item{dpi}{Input the dpi of the image to create}

\item{width}{Input the width of the image to create}

\item{cmpdNm}{Input the name of the compound to plot}
}
\description{
Plot compound summary for multi-linear regression tool
}
\author{
Jessica Ewald\email{jessica.ewald@mcgill.ca}
McGill University, Canada
License: GPL-3 License
}
