% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_qc_utils.R
\name{PlotPhosphoQC}
\alias{PlotPhosphoQC}
\title{Plot QC panel (quantify) for phosphoproteomics data (disabled)}
\usage{
PlotPhosphoQC(fileName, imgNm, dpi = 96, format = "png", panel = "quantify")
}
\description{
Plot QC panel (quantify) for phosphoproteomics data (disabled)
}
