% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotRocUnivBoxPlot}
\alias{PlotRocUnivBoxPlot}
\title{Plot a boxplot view of a selected feature}
\usage{
PlotRocUnivBoxPlot(
  dataName = "",
  feat.nm,
  version,
  format = "png",
  dpi = 72,
  isOpt,
  isQuery,
  dataType = "default"
)
}
\arguments{
\item{dataName}{Input dataset name}

\item{feat.nm}{Input the feature name}

\item{version}{version mark for image name}

\item{format}{Select the image format}

\item{dpi}{Input the dpi}

\item{isOpt}{logical}

\item{isQuery}{logical}
}
\description{
Plots a boxplot of the selected feature's concentration
}
