% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Predict.class}
\alias{Predict.class}
\title{Predict class probabilities}
\usage{
Predict.class(x.train, y.train, x.test, clsMethod = "pls", lvNum, imp.out = F)
}
\arguments{
\item{x.train}{Training features}

\item{y.train}{Training labels}

\item{x.test}{Test features}

\item{clsMethod}{Classification method}

\item{lvNum}{Number of components}

\item{imp.out}{Return feature importances?}
}
\description{
Predict class probabilities
}
