% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_network_utils.R
\name{PrepareNetwork}
\alias{PrepareNetwork}
\title{PrepareNetwork}
\usage{
PrepareNetwork(net.nm, json.nm)
}
\arguments{
\item{net.nm}{Name of the network (e.g., "module1", "ppi_1")}

\item{json.nm}{Output JSON file name}
}
\value{
1 if successful, 0 otherwise
}
\description{
Exports a PPI network to JSON format for visualization
This is the main entry point called from Java for network viewer
}
