% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PrepareROCData}
\alias{PrepareROCData}
\title{Prepare data for ROC analysis}
\usage{
PrepareROCData(dataName = "", sel.meta = "NA", factor1 = "NA", factor2 = "NA")
}
\arguments{
\item{dataName}{Input dataset name}

\item{sel.meta}{Selected metadata column}

\item{factor1}{First factor level}

\item{factor2}{Second factor level}
}
\description{
Prepare data for ROC analysis with metadata selection
}
