% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PrepareROCData_old}
\alias{PrepareROCData_old}
\title{Prepare ROC data (old version)}
\usage{
PrepareROCData_old(dataName = "")
}
\arguments{
\item{dataName}{Input dataset name}
}
\description{
Prepare data for ROC analysis without metadata
}
