% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PrepareROCDetails}
\alias{PrepareROCDetails}
\title{ROC with CI for AUC}
\usage{
PrepareROCDetails(dataName = "", feat.nm)
}
\arguments{
\item{dataName}{Input dataset name}

\item{feat.nm}{Input the feature name}
}
\description{
Prepare detailed ROC analysis for a feature
}
