% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_utils.R
\name{PrepareUpsetDataFromComparisons}
\alias{PrepareUpsetDataFromComparisons}
\title{Prepare data for Upset diagram using DE comparisons (onedata)}
\usage{
PrepareUpsetDataFromComparisons(fileNm, dataName = "")
}
\arguments{
\item{fileNm}{file name of the json file output}

\item{dataName}{optional dataset name (defaults to paramSet$dataName)}
}
\description{
Prepare data for Upset diagram using DE comparisons (onedata)
}
