% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_qc_utils.R
\name{qc.loc.prob.hist}
\alias{qc.loc.prob.hist}
\title{Internal function: Localization probability histogram
Runs in quiet mode (backend) without triggering Quartz/GUI}
\usage{
qc.loc.prob.hist(loc_probs, imgNm, dpi = 96, format = "png")
}
\description{
Internal function: Localization probability histogram
Runs in quiet mode (backend) without triggering Quartz/GUI
}
