% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_qc_utils.R
\name{qc.protein.cv.hist}
\alias{qc.protein.cv.hist}
\title{Internal function: Protein CV Distribution Histogram}
\usage{
qc.protein.cv.hist(data_mat, groups, imgNm, dpi = 96, format = "png")
}
\description{
Internal function: Protein CV Distribution Histogram
}
