% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_graphics.R
\name{qc.sample.corr}
\alias{qc.sample.corr}
\title{Sample correlation heatmap (metadata-style aesthetic)}
\usage{
qc.sample.corr(
  dat,
  imgNm,
  dpi = 96,
  format = "png",
  interactive = FALSE,
  method = "pearson",
  colorGradient = "default"
)
}
\description{
Sample correlation heatmap (metadata-style aesthetic)
}
