% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{RankFeatures}
\alias{RankFeatures}
\title{Rank features using different methods}
\usage{
RankFeatures(x.in, y.in, method, lvNum)
}
\arguments{
\item{x.in}{Feature matrix (samples × features)}

\item{y.in}{Class labels}

\item{method}{Ranking method: "auroc", "ttest", "foldchange"}

\item{lvNum}{Number of latent variables for PLS}
}
\description{
Rank features using different methods
}
