% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readtable_utils.R
\name{ReadMetaData}
\alias{ReadMetaData}
\title{Read Metadata for Meta-Analysis Mode}
\usage{
ReadMetaData(metafilename)
}
\arguments{
\item{metafilename}{A character string specifying the name of the metadata file.}
}
\value{
An integer indicating the success of reading and processing metadata.
}
\description{
This function reads metadata for meta-analysis mode and performs necessary checks.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
ReadMetaData(metafilename = "metadata.csv")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
