% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{ReadOmicsData}
\alias{ReadOmicsData}
\title{Read Omics Data for Meta-Analysis}
\usage{
ReadOmicsData(fileName)
}
\arguments{
\item{fileName}{File name of the data table.}
}
\value{
A registered dataset object for further analysis.
}
\description{
This function reads gene expression data and parses it into an R object for meta-analysis.
}
\details{
McGill University, Canada
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
}
