% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_readtable_utils.R
\name{ReadTabExpressData}
\alias{ReadTabExpressData}
\title{Read Tabular Expression Data and Metadata}
\usage{
ReadTabExpressData(
  fileName,
  metafileName = "",
  metaContain = "true",
  oneDataAnalType = "default",
  path = "",
  dataFormat = "default"
)
}
\arguments{
\item{fileName}{A character string specifying the name of the expression data file.}

\item{metafileName}{A character string specifying the name of the metadata file, ignore if metaContain = true}

\item{metaContain}{A logical value indicating whether metadata is contained in the data file if false, metadata is in separate file}

\item{oneDataAnalType}{The type of analysis to perform on the one-data setup.}

\item{path}{The path to the files if they are located in a different directory.}
}
\value{
A processed dataset object containing expression data and metadata.
}
\description{
This function reads tabular expression data along with metadata and processes the data.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
ReadTabExpressData(fileName = "expression_data.csv", metafileName = "metadata.csv",
                   metaContain = TRUE, oneDataAnalType = "default", path = "")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
