% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{RemoveMissingPercent}
\alias{RemoveMissingPercent}
\title{Remove Variables with High Missing Percentage}
\usage{
RemoveMissingPercent(dataName = "", percent = perct)
}
\arguments{
\item{dataName}{The name of the dataset to be processed.}

\item{percent}{The threshold percentage of missing values above which variables will be removed.}
}
\description{
This function removes variables (columns) from the dataset that have a high percentage of missing values.
}
\details{
Additional details about the function, if needed.
}
\examples{
\dontrun{
RemoveMissingPercent("data_filename", 0.2)
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
