% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils_meta.R
\name{RestoreMetaData}
\alias{RestoreMetaData}
\title{Restore raw metadata (undo ComBat)}
\usage{
RestoreMetaData()
}
\value{
1 if success, 0 otherwise.
}
\description{
Reloads the original metadata that was stored before batch correction.
}
