% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_utils.R
\name{run_func_via_rsclient}
\alias{run_func_via_rsclient}
\title{Execute function in isolated RSclient fork}
\usage{
run_func_via_rsclient(func, args = list(), timeout_sec = 60)
}
\arguments{
\item{func}{Function to run in forked Rserve child}

\item{args}{List of arguments passed via do.call}

\item{timeout_sec}{Hard timeout before child is killed}
}
\value{
Result of do.call(func, args)
}
\description{
Execute function in isolated RSclient fork
}
