% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{SampleNorm_Median}
\alias{SampleNorm_Median}
\title{Normalization to sample median
Each sample is divided by its median value}
\usage{
SampleNorm_Median(data)
}
\arguments{
\item{data}{Matrix with samples as columns}
}
\value{
Normalized matrix
}
\description{
Normalization to sample median
Each sample is divided by its median value
}
