% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{SampleNorm_Sum}
\alias{SampleNorm_Sum}
\title{Normalization to constant sum
Each sample is divided by its sum and multiplied by a constant (1000)}
\usage{
SampleNorm_Sum(data)
}
\arguments{
\item{data}{Matrix with samples as columns}
}
\value{
Normalized matrix
}
\description{
Normalization to constant sum
Each sample is divided by its sum and multiplied by a constant (1000)
}
