% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{SanityCheckData}
\alias{SanityCheckData}
\title{Perform Data Integrity Check}
\usage{
SanityCheckData(fileName)
}
\arguments{
\item{fileName}{The name of the data file to be checked.}
}
\value{
A modified dataset after performing data sanity checks.
}
\description{
This function performs a sanity check on the input data, including class labels and data matrix.
}
\examples{
\dontrun{
SanityCheckData("data_filename")
}

}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
