% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_proc.R
\name{SanityCheckMetaData}
\alias{SanityCheckMetaData}
\title{Perform Integrity Check on Metadata}
\usage{
SanityCheckMetaData()
}
\value{
integer 0 or 1 to indicate success or failure
}
\description{
This function performs a sanity check on the metadata associated with the datasets.
}
\author{
Guangyan Zhou \email{guangyan.zhou@mail.mcgill.ca}
}
