% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmap_wrapper_utils.R
\name{SaveHeatmapJSON}
\alias{SaveHeatmapJSON}
\title{Prepare data for heatmap visualization}
\usage{
SaveHeatmapJSON(dataName = "", fileName, displayOpt)
}
\description{
Prepare data for heatmap visualization
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: MIT
}
