% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{SetAnalysisMode}
\alias{SetAnalysisMode}
\title{Set analysis mode}
\usage{
SetAnalysisMode(dataName = "", mode)
}
\arguments{
\item{dataName}{Dataset name}

\item{mode}{Analysis mode: "univ" or "multivar"}
}
\description{
Set analysis mode
}
