% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{SetCurrentGroups}
\alias{SetCurrentGroups}
\title{Choose two groups (when more than two groups uploaded)}
\usage{
SetCurrentGroups(dataName = "", grps)
}
\arguments{
\item{dataName}{Input the dataset name}

\item{grps}{Group names to compare}
}
\description{
Choose two groups (when more than two groups uploaded)
}
