% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{SetCustomData}
\alias{SetCustomData}
\title{Set custom data for ROC}
\usage{
SetCustomData(dataName = "", selected.cmpds, selected.smpls)
}
\arguments{
\item{dataName}{Input dataset name}

\item{selected.cmpds}{Vector of selected compounds}

\item{selected.smpls}{Vector of selected samples}
}
\description{
Set custom data selection for hold-out samples
}
