% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_deanal_utils.R
\name{SetSelectedMetaInfo}
\alias{SetSelectedMetaInfo}
\title{Select Metadata Factors for DE Analysis}
\usage{
SetSelectedMetaInfo(dataName = "", meta0, meta1, block1)
}
\arguments{
\item{dataName}{A character string specifying the name of the dataset.}

\item{meta0}{The primary metadata factor}

\item{meta1}{The secondary metadata factor}

\item{block1}{The blocking factor for}
}
\description{
This function is used to select metadata factors for a Differential Expression (DE) analysis.
}
\details{
Additional details about the function, if needed.
}
\author{
Guangyan Zhou, \email{guangyan.zhou@mail.mcgill.ca}
}
