% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{UpdateKmeans}
\alias{UpdateKmeans}
\title{Update k-means clustering}
\usage{
UpdateKmeans(dataName = "", clust.num = 5)
}
\arguments{
\item{dataName}{Dataset name}

\item{clust.num}{Number of clusters}
}
\description{
Update k-means clustering
}
