% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_data_utils.R
\name{.parseDIANNModifiedSequence}
\alias{.parseDIANNModifiedSequence}
\title{.parseDIANNModifiedSequence}
\usage{
.parseDIANNModifiedSequence(mod_seq, protein_ids, stripped_seq)
}
\arguments{
\item{mod_seq}{Vector of modified sequences}

\item{protein_ids}{Vector of protein IDs}

\item{stripped_seq}{Vector of stripped sequences (no modifications)}
}
\value{
Data frame with columns: site_id, protein_id, residue, position
}
\description{
Parse DIA-NN Modified.Sequence to extract phosphosite positions
}
\details{
DIA-NN format examples:
"PEPT(Phospho (STY))IDE"  -> Phospho on T at position 4
"PEPS(UniMod:21)EQRES(UniMod:21)" -> Two phosphosites (UniMod:21 = Phospho)
"M(Oxidation (M))PEPTIDE" -> Oxidation (not phospho, skip)
"(UniMod:1)PEPS(UniMod:21)TIDE" -> N-term acetyl + phospho
}
