% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phospho_enrich_utils.R
\name{PerformKinaseEnrichment}
\alias{PerformKinaseEnrichment}
\title{Perform kinase enrichment analysis on significant phosphosites}
\usage{
PerformKinaseEnrichment(dataName, database = "phosphositeplus")
}
\arguments{
\item{dataName}{Dataset name}

\item{database}{Kinase-substrate database to use (matches files in resources/data/ksea, or "all")}
}
\value{
1 on success, 0 on failure
}
\description{
Perform kinase enrichment analysis on significant phosphosites
}
