% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{Plot.Permutation}
\alias{Plot.Permutation}
\title{Plot results of permutation tests}
\usage{
Plot.Permutation(mSetObj=NA, imgName, format="png", dpi=default.dpi)
}
\arguments{
\item{imgName}{Input a name for the plot}

\item{format}{elect the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images,
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Plot results of permutation tests
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
