% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotAccuracy}
\alias{PlotAccuracy}
\title{Plot classification performance using different features for Multi-Biomarker}
\usage{
PlotAccuracy(mSetObj=NA, imgName, format="png", dpi=default.dpi)
}
\arguments{
\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images, 
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{mSetObj}{Input the name of the created mSetObj (see InitDataObjects)}
}
\description{
Plot of the accuracy of classification with an increasing number of features.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
