% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomarker_utils.R
\name{PlotProbViewTest}
\alias{PlotProbViewTest}
\title{Plot a summary view of the classification result of tester prediction}
\usage{
PlotProbViewTest(dataName = "", imgName, format="png", dpi=default.dpi, mdl.inx, show, showPred)
}
\arguments{
\item{imgName}{Input a name for the plot}

\item{format}{Select the image format, "png", of "pdf".}

\item{dpi}{Input the dpi. If the image format is "pdf", users need not define the dpi. For "png" images,
the default dpi is 72. It is suggested that for high-resolution images, select a dpi of 300.}

\item{mdl.inx}{Model index, 0 means to compare all models, -1 means to use the best model, input 1-6 to plot a ROC curve for one of the top six models}

\item{show}{1 or 0, if 1, label samples classified to the wrong groups}

\item{showPred}{Show predicted samples}

\item{dataName}{Input the name of the dataset (see \code{Init.Data})}
}
\description{
Plot of predicted class probabilities. On the x-axis is the proability,
and the y-axis is the index of each predicted sample based on the probility.
The samples are turned into separations at the x-axis.
This plot can be created for multivariate ROC curve analysis using SVM, PLS, and RandomForest.
Please note that sometimes, not all samples will be tested, instead they will be plotted
at the 0.5 neutral line.
}
\author{
Jeff Xia \email{jeff.xia@mcgill.ca}
McGill University, Canada
License: GNU GPL (>= 2)
}
