% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_utils.R
\name{SampleNorm_PQN}
\alias{SampleNorm_PQN}
\title{Probabilistic Quotient Normalization (PQN)
\enumerate{
\item Perform an integral normalization (sum normalization)
\item Choose/calculate a reference spectrum (median across samples)
\item Calculate the quotients of all variables of interest relative to the reference
\item Calculate the median of these quotients for each sample
\item Divide all variables of each sample by this median quotient
}}
\usage{
SampleNorm_PQN(data)
}
\arguments{
\item{data}{Matrix with samples as columns}
}
\value{
Normalized matrix
}
\description{
Probabilistic Quotient Normalization (PQN)
\enumerate{
\item Perform an integral normalization (sum normalization)
\item Choose/calculate a reference spectrum (median across samples)
\item Calculate the quotients of all variables of interest relative to the reference
\item Calculate the median of these quotients for each sample
\item Divide all variables of each sample by this median quotient
}
}
